backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pymysql>=1.0.2
python-dotenv>=0.19.0
pydantic>=1.8.2
aiomysql>=0.0.22
python-multipart>=0.0.5